Cut down slide panels, added magnify
