+++ src/components/overview/ImageCarousel.jsx
@@ -1,7 +1,8 @@
-import React, { useState, useEffect } from "react";
-import { createCloudinaryDisplayURL } from "/src/services/Cloudinary.js";
+import React, { useState, useEffect, useRef } from "react";
+import { createCloudinaryDisplayURL } from "../../services/Cloudinary.js";
 import ThumbnailContainer from "./ThumbnailContainer.jsx";
 import ExpandedProductInfo from "./ExpandedProductInfo.jsx";
+import CollapseButton from "./CollapseButton.jsx";
 
 const ImageCarousel = ({
   image,
@@ -20,8 +21,14 @@ const ImageCarousel = ({
   decrementThumbnailIndex,
   rating,
 }) => {
+  let [collapsePanel, setCollapsePanel] = useState(false);
+
+  let callHiding = function (callback) {
+    callback();
+  };
+
   return (
-    <div className="overview-image-container" data-testid="imageCarousel">
+    <div className="overview-image-container" data-testid="image-carousel">
       <button
         onClick={() => {
           decrementThumbnailIndex();
@@ -40,6 +47,9 @@ const ImageCarousel = ({
         photos={photoObjects}
         activeThumbnailIndex={activeThumbnailIndex}></ThumbnailContainer>
       <ExpandedProductInfo
+        collapsePanel={collapsePanel}
+        setCollapsePanel={setCollapsePanel}
+        callHiding={callHiding}
         rating={rating}
         incrementThumbnailIndex={incrementThumbnailIndex}
         activeThumbnailIndex={activeThumbnailIndex}
@@ -52,6 +62,10 @@ const ImageCarousel = ({
         styleInfo={styleInfo}
         productInfo={productInfo}
         styleObjects={styleObjects}></ExpandedProductInfo>
+      <CollapseButton
+        setIsHiding={() => {
+          setCollapsePanel(!collapsePanel);
+        }}></CollapseButton>
     </div>
   );
 };
