+++ src/components/overview/PurchaseInfo.test.js
