+++ src/components/overview/ThumbnailContainer.test.js
