+++ src/components/overviewOld/ImageCarousel.jsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from "react";
-import { createCloudinaryDisplayURL } from "../../services/Cloudinary.js";
+import { createCloudinaryDisplayURL } from "/src/services/Cloudinary.js";
 import ThumbnailContainer from "./ThumbnailContainer.jsx";
 import ExpandedProductInfo from "./ExpandedProductInfo.jsx";
 
@@ -21,7 +21,7 @@ const ImageCarousel = ({
   rating,
 }) => {
   return (
-    <div className="overview-image-container" data-testid="image-carousel">
+    <div className="overview-image-container" data-testid="imageCarousel">
       <button
         onClick={() => {
           decrementThumbnailIndex();
