+++ src/components/overviewOld/Price.jsx
@@ -4,18 +4,16 @@ const Price = ({ styleInfo }) => {
   if (styleInfo.sale_price) {
     return (
       <div data-testid="price1" className="overview-expanded-product-info-prices">
-        <span data-testid="original-price-struck" className="overview-expanded-product-info-price-struck">
+        <span data-testid="test-price-struck" className="overview-expanded-product-info-price-struck">
           {styleInfo.original_price}
         </span>
-        <span data-testid="sale-price" className="overview-expanded-product-info-different-price">
-          {styleInfo.sale_price}
-        </span>
+        <span className="overview-expanded-product-info-different-price">{styleInfo.sale_price}</span>
       </div>
     );
   } else {
     return (
       <div data-testid="price2" className="overview-expanded-product-info-prices">
-        <span data-testid="original-price">{styleInfo.original_price}</span>
+        <span>{styleInfo.original_price}</span>
       </div>
     );
   }
