+++ src/components/overviewOld/ThumbnailDecrement.jsx
