+++ src/index.js
@@ -3,6 +3,7 @@ import { createRoot } from "react-dom/client";
 import App from "./App";
 import "./assets/fontAwesome/js/all.min.js";
 import "./assets/fontAwesome/css/all.min.css";
+import "/public/";
 
 // reactDom.render(<App />, document.getElementById("root"));
